requirements.txt
av==14.0.1
dataclasses_json==0.6.7
decord==0.6.0
Flask==3.1.0
flask_cors==5.0.0
huggingface_hub==0.27.1
hydra-core==1.3.2
imagesize==1.4.1
iopath==0.1.10
matplotlib==3.7.2
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.15.1
opencv_python==4.8.0.76
pandas==2.0.3
Pillow==11.1.0
pycocotools==2.0.7
PyQt5==5.15.11
PyQt5_sip==12.13.0
setuptools==68.0.0
scikit-image
strawberry==3.0
submitit==1.5.2
tensordict==0.6.2
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.4
